Fixed exception handling exception

+++ CryoCloud/Common/fileprep.py
@@ -101,7 +101,9 @@ class FilePrepare:
         except Exception as e:
             # retval["errors"] += "%s: %s\n" % (s, e)
             self.log.exception("Unzip of %s failed: %s" % (s, str(e)))
-            shutil.rmtree(decomp)
+            if os.path.exists(decomp):
+                shutil.rmtree(decomp)
+            raise e
 
         # We now rename the temporary directory to the destination name
         try:
